Put 'patch_api_header.rb' in separate folder and enhance 'ns1.rb'
