+++ test/fixtures/vcr_cassettes/ns1_add_changeset.yml
@@ -21,7 +21,7 @@ http_interactions:
       message: Not Found
     headers:
       Date:
-      - Sat, 09 Nov 2019 17:29:17 GMT
+      - Mon, 11 Nov 2019 18:00:04 GMT
       Content-Type:
       - application/json
       Content-Length:
@@ -29,8 +29,8 @@ http_interactions:
       Connection:
       - keep-alive
       Set-Cookie:
-      - __cfduid=d2a0cb6bfe03c685c4039d98786e614b11573320557; expires=Sun, 08-Nov-20
-        17:29:17 GMT; path=/; domain=.nsone.net; HttpOnly
+      - __cfduid=d4c7def9cd5c2d7cca937437a04c0612c1573495204; expires=Tue, 10-Nov-20
+        18:00:04 GMT; path=/; domain=.nsone.net; HttpOnly
       X-Ratelimit-Remaining:
       - '897'
       X-Ratelimit-By:
@@ -52,14 +52,14 @@ http_interactions:
       Server:
       - cloudflare
       Cf-Ray:
-      - 53318209efe427d2-YYZ
+      - 534229e14f4bab4c-YYZ
     body:
       encoding: UTF-8
       string: '{"message":"record not found"}
 
         '
     http_version: 
-  recorded_at: Sat, 09 Nov 2019 17:29:17 GMT
+  recorded_at: Mon, 11 Nov 2019 18:00:04 GMT
 - request:
     method: put
     uri: https://api.nsone.net/v1/zones/test.recordstore.io/test_add_changeset.test.recordstore.io/A
@@ -83,7 +83,7 @@ http_interactions:
       message: OK
     headers:
       Date:
-      - Sat, 09 Nov 2019 17:29:17 GMT
+      - Mon, 11 Nov 2019 18:00:04 GMT
       Content-Type:
       - application/json
       Transfer-Encoding:
@@ -91,10 +91,10 @@ http_interactions:
       Connection:
       - keep-alive
       Set-Cookie:
-      - __cfduid=d1ddb328c19bfc3fc0dfb36f78f140d141573320557; expires=Sun, 08-Nov-20
-        17:29:17 GMT; path=/; domain=.nsone.net; HttpOnly
+      - __cfduid=d77bae45be66b1694a8bd140a4d6355d31573495204; expires=Tue, 10-Nov-20
+        18:00:04 GMT; path=/; domain=.nsone.net; HttpOnly
       X-Ratelimit-Remaining:
-      - '98'
+      - '99'
       X-Ratelimit-By:
       - customer
       X-Ratelimit-Limit:
@@ -123,12 +123,12 @@ http_interactions:
       Server:
       - cloudflare
       Cf-Ray:
-      - 5331820cafe0cab8-YYZ
+      - 534229e25ffc27ea-YYZ
     body:
       encoding: ASCII-8BIT
-      string: '{"domain":"test_add_changeset.test.recordstore.io","zone":"test.recordstore.io","use_client_subnet":true,"answers":[{"answer":["10.10.10.42"],"id":"5dc6f76dc9c79d0001a62e58"}],"id":"5dc6f76dc9c79d0001a62e59","regions":{},"meta":{},"link":null,"filters":[],"ttl":600,"tier":1,"type":"A","networks":[0]}
+      string: '{"domain":"test_add_changeset.test.recordstore.io","zone":"test.recordstore.io","use_client_subnet":true,"answers":[{"answer":["10.10.10.42"],"id":"5dc9a1a4c94a9000011b3c22"}],"id":"5dc9a1a4c94a9000011b3c23","regions":{},"meta":{},"link":null,"filters":[],"ttl":600,"tier":1,"type":"A","networks":[0]}
 
         '
     http_version: 
-  recorded_at: Sat, 09 Nov 2019 17:29:17 GMT
+  recorded_at: Mon, 11 Nov 2019 18:00:05 GMT
 recorded_with: VCR 5.0.0
