+++ test/fixtures/vcr_cassettes/ns1_add_changeset_missing_zone.yml
@@ -21,7 +21,7 @@ http_interactions:
       message: Bad Request
     headers:
       Date:
-      - Sat, 09 Nov 2019 17:29:16 GMT
+      - Mon, 11 Nov 2019 18:00:25 GMT
       Content-Type:
       - application/json
       Content-Length:
@@ -29,10 +29,10 @@ http_interactions:
       Connection:
       - keep-alive
       Set-Cookie:
-      - __cfduid=df713beb0340fed82106eb1a998c793681573320556; expires=Sun, 08-Nov-20
-        17:29:16 GMT; path=/; domain=.nsone.net; HttpOnly
+      - __cfduid=d36fc80835ca7e643f9fcdc8d6d475b551573495225; expires=Tue, 10-Nov-20
+        18:00:25 GMT; path=/; domain=.nsone.net; HttpOnly
       X-Ratelimit-Remaining:
-      - '896'
+      - '893'
       X-Ratelimit-By:
       - customer
       X-Ratelimit-Limit:
@@ -52,7 +52,7 @@ http_interactions:
       Server:
       - cloudflare
       Cf-Ray:
-      - 533182076ea1abac-YYZ
+      - 53422a687eeeab7c-YYZ
     body:
       encoding: UTF-8
       string: '{"message":"\"test_add_changset_missing_zone.test.recordstore.io\"
@@ -60,7 +60,7 @@ http_interactions:
 
         '
     http_version: 
-  recorded_at: Sat, 09 Nov 2019 17:29:16 GMT
+  recorded_at: Mon, 11 Nov 2019 18:00:25 GMT
 - request:
     method: put
     uri: https://api.nsone.net/v1/zones/test_add_changset_missing_zone.recordstore.io/test_add_changset_missing_zone.test.recordstore.io/A
@@ -84,7 +84,7 @@ http_interactions:
       message: Bad Request
     headers:
       Date:
-      - Sat, 09 Nov 2019 17:29:16 GMT
+      - Mon, 11 Nov 2019 18:00:25 GMT
       Content-Type:
       - application/json
       Content-Length:
@@ -92,8 +92,8 @@ http_interactions:
       Connection:
       - keep-alive
       Set-Cookie:
-      - __cfduid=d28d6120d666f31b87d7392f58df078c51573320556; expires=Sun, 08-Nov-20
-        17:29:16 GMT; path=/; domain=.nsone.net; HttpOnly
+      - __cfduid=d8a4292b6a538e800558d904003de51901573495225; expires=Tue, 10-Nov-20
+        18:00:25 GMT; path=/; domain=.nsone.net; HttpOnly
       X-Ratelimit-Remaining:
       - '99'
       X-Ratelimit-By:
@@ -115,7 +115,7 @@ http_interactions:
       Server:
       - cloudflare
       Cf-Ray:
-      - 533182086fe5ab4c-YYZ
+      - 53422a696b62ab94-YYZ
     body:
       encoding: UTF-8
       string: '{"message":"\"test_add_changset_missing_zone.test.recordstore.io\"
@@ -123,5 +123,5 @@ http_interactions:
 
         '
     http_version: 
-  recorded_at: Sat, 09 Nov 2019 17:29:16 GMT
+  recorded_at: Mon, 11 Nov 2019 18:00:25 GMT
 recorded_with: VCR 5.0.0
