+++ test/fixtures/vcr_cassettes/ns1_add_multiple_changesets.yml
@@ -21,7 +21,7 @@ http_interactions:
       message: Not Found
     headers:
       Date:
-      - Sat, 09 Nov 2019 17:29:25 GMT
+      - Mon, 11 Nov 2019 17:59:56 GMT
       Content-Type:
       - application/json
       Content-Length:
@@ -29,10 +29,10 @@ http_interactions:
       Connection:
       - keep-alive
       Set-Cookie:
-      - __cfduid=d89852fe253dc4738f73218dae77249dc1573320565; expires=Sun, 08-Nov-20
-        17:29:25 GMT; path=/; domain=.nsone.net; HttpOnly
+      - __cfduid=d0be4914fe95a596d462ca005f76665e61573495196; expires=Tue, 10-Nov-20
+        17:59:56 GMT; path=/; domain=.nsone.net; HttpOnly
       X-Ratelimit-Remaining:
-      - '886'
+      - '899'
       X-Ratelimit-By:
       - customer
       X-Ratelimit-Limit:
@@ -52,14 +52,14 @@ http_interactions:
       Server:
       - cloudflare
       Cf-Ray:
-      - 5331823c3f7bcab4-YYZ
+      - 534229b3d97fcab8-YYZ
     body:
       encoding: UTF-8
       string: '{"message":"record not found"}
 
         '
     http_version: 
-  recorded_at: Sat, 09 Nov 2019 17:29:25 GMT
+  recorded_at: Mon, 11 Nov 2019 17:59:56 GMT
 - request:
     method: put
     uri: https://api.nsone.net/v1/zones/test.recordstore.io/test_add_multiple_changesets.test.recordstore.io/A
@@ -83,7 +83,7 @@ http_interactions:
       message: OK
     headers:
       Date:
-      - Sat, 09 Nov 2019 17:29:25 GMT
+      - Mon, 11 Nov 2019 17:59:57 GMT
       Content-Type:
       - application/json
       Transfer-Encoding:
@@ -91,10 +91,10 @@ http_interactions:
       Connection:
       - keep-alive
       Set-Cookie:
-      - __cfduid=d65b19d8c34b7b4063d5ef163191cdee31573320565; expires=Sun, 08-Nov-20
-        17:29:25 GMT; path=/; domain=.nsone.net; HttpOnly
+      - __cfduid=d1a1b7b2ed837b6734171707b264045931573495196; expires=Tue, 10-Nov-20
+        17:59:56 GMT; path=/; domain=.nsone.net; HttpOnly
       X-Ratelimit-Remaining:
-      - '97'
+      - '99'
       X-Ratelimit-By:
       - customer
       X-Ratelimit-Limit:
@@ -123,14 +123,14 @@ http_interactions:
       Server:
       - cloudflare
       Cf-Ray:
-      - 5331823d6edbabac-YYZ
+      - 534229b4ff11aba6-YYZ
     body:
       encoding: ASCII-8BIT
-      string: '{"domain":"test_add_multiple_changesets.test.recordstore.io","zone":"test.recordstore.io","use_client_subnet":true,"answers":[{"answer":["10.10.10.42"],"id":"5dc6f775c94a900001011184"}],"id":"5dc6f775c94a900001011185","regions":{},"meta":{},"link":null,"filters":[],"ttl":600,"tier":1,"type":"A","networks":[0]}
+      string: '{"domain":"test_add_multiple_changesets.test.recordstore.io","zone":"test.recordstore.io","use_client_subnet":true,"answers":[{"answer":["10.10.10.42"],"id":"5dc9a19cc9c79d0001326ac1"}],"id":"5dc9a19cc9c79d0001326ac2","regions":{},"meta":{},"link":null,"filters":[],"ttl":600,"tier":1,"type":"A","networks":[0]}
 
         '
     http_version: 
-  recorded_at: Sat, 09 Nov 2019 17:29:25 GMT
+  recorded_at: Mon, 11 Nov 2019 17:59:57 GMT
 - request:
     method: get
     uri: https://api.nsone.net/v1/zones/test.recordstore.io/test_add_multiple_changesets.test.recordstore.io/A
@@ -152,7 +152,7 @@ http_interactions:
       message: OK
     headers:
       Date:
-      - Sat, 09 Nov 2019 17:29:25 GMT
+      - Mon, 11 Nov 2019 17:59:59 GMT
       Content-Type:
       - application/json
       Transfer-Encoding:
@@ -160,12 +160,12 @@ http_interactions:
       Connection:
       - keep-alive
       Set-Cookie:
-      - __cfduid=d7dc19fc54765cb4b50a012313fbb204a1573320565; expires=Sun, 08-Nov-20
-        17:29:25 GMT; path=/; domain=.nsone.net; HttpOnly
+      - __cfduid=df8bac13a7ee97189e580a93e4f877a7c1573495199; expires=Tue, 10-Nov-20
+        17:59:59 GMT; path=/; domain=.nsone.net; HttpOnly
       X-Ratelimit-Remaining:
-      - '887'
+      - '899'
       Etag:
-      - W/"0150a07943981d9207c7288f52a1e1e2369cf557"
+      - W/"d1500566f489cb055343a2b6f4efa1858a0e90ac"
       X-Ratelimit-By:
       - customer
       X-Ratelimit-Limit:
@@ -194,20 +194,20 @@ http_interactions:
       Server:
       - cloudflare
       Cf-Ray:
-      - 5331823f6fc23010-YYZ
+      - 534229c28dcbab82-YYZ
     body:
       encoding: ASCII-8BIT
-      string: '{"domain":"test_add_multiple_changesets.test.recordstore.io","zone":"test.recordstore.io","use_client_subnet":true,"answers":[{"answer":["10.10.10.42"],"id":"5dc6f775c94a900001011184"}],"id":"5dc6f775c94a900001011185","regions":{},"meta":{},"link":null,"filters":[],"ttl":600,"tier":1,"type":"A","networks":[0]}
+      string: '{"domain":"test_add_multiple_changesets.test.recordstore.io","zone":"test.recordstore.io","use_client_subnet":true,"answers":[{"answer":["10.10.10.42"],"id":"5dc9a19cc9c79d0001326ac1"}],"id":"5dc9a19cc9c79d0001326ac2","regions":{},"meta":{},"link":null,"filters":[],"ttl":600,"tier":1,"type":"A","networks":[0]}
 
         '
     http_version: 
-  recorded_at: Sat, 09 Nov 2019 17:29:25 GMT
+  recorded_at: Mon, 11 Nov 2019 17:59:59 GMT
 - request:
     method: post
     uri: https://api.nsone.net/v1/zones/test.recordstore.io/test_add_multiple_changesets.test.recordstore.io/A
     body:
       encoding: UTF-8
-      string: '{"answers":[{"answer":["10.10.10.42"],"id":"5dc6f775c94a900001011184"},{"answer":["10.10.10.43"]}],"ttl":600}'
+      string: '{"answers":[{"answer":["10.10.10.42"],"id":"5dc9a19cc9c79d0001326ac1"},{"answer":["10.10.10.43"]}],"ttl":600}'
     headers:
       X-Nsone-Key:
       - "<NS1_API_KEY>"
@@ -225,7 +225,7 @@ http_interactions:
       message: OK
     headers:
       Date:
-      - Sat, 09 Nov 2019 17:29:26 GMT
+      - Mon, 11 Nov 2019 17:59:59 GMT
       Content-Type:
       - application/json
       Transfer-Encoding:
@@ -233,8 +233,8 @@ http_interactions:
       Connection:
       - keep-alive
       Set-Cookie:
-      - __cfduid=d9cfd53f2ef8820f510a78409a7aa39f91573320565; expires=Sun, 08-Nov-20
-        17:29:25 GMT; path=/; domain=.nsone.net; HttpOnly
+      - __cfduid=d3244c164f99bed42de55d0dd461044b11573495199; expires=Tue, 10-Nov-20
+        17:59:59 GMT; path=/; domain=.nsone.net; HttpOnly
       X-Ratelimit-Remaining:
       - '299'
       X-Ratelimit-By:
@@ -265,12 +265,12 @@ http_interactions:
       Server:
       - cloudflare
       Cf-Ray:
-      - 533182405f60ca9c-YYZ
+      - 534229c38a963022-YYZ
     body:
       encoding: ASCII-8BIT
-      string: '{"domain":"test_add_multiple_changesets.test.recordstore.io","zone":"test.recordstore.io","use_client_subnet":true,"answers":[{"answer":["10.10.10.42"],"id":"5dc6f775c94a900001011184"},{"answer":["10.10.10.43"],"id":"5dc6f776c94a9000018788ae"}],"id":"5dc6f775c94a900001011185","regions":{},"meta":{},"link":null,"filters":[],"ttl":600,"tier":1,"type":"A","networks":[0]}
+      string: '{"domain":"test_add_multiple_changesets.test.recordstore.io","zone":"test.recordstore.io","use_client_subnet":true,"answers":[{"answer":["10.10.10.42"],"id":"5dc9a19cc9c79d0001326ac1"},{"answer":["10.10.10.43"],"id":"5dc9a19fc9c79d000145db5b"}],"id":"5dc9a19cc9c79d0001326ac2","regions":{},"meta":{},"link":null,"filters":[],"ttl":600,"tier":1,"type":"A","networks":[0]}
 
         '
     http_version: 
-  recorded_at: Sat, 09 Nov 2019 17:29:26 GMT
+  recorded_at: Mon, 11 Nov 2019 17:59:59 GMT
 recorded_with: VCR 5.0.0
