+++ test/fixtures/vcr_cassettes/ns1_get_zones.yml
@@ -21,7 +21,7 @@ http_interactions:
       message: OK
     headers:
       Date:
-      - Sat, 09 Nov 2019 17:29:26 GMT
+      - Mon, 11 Nov 2019 18:00:53 GMT
       Content-Type:
       - application/json
       Transfer-Encoding:
@@ -29,12 +29,12 @@ http_interactions:
       Connection:
       - keep-alive
       Set-Cookie:
-      - __cfduid=dccc792ad6d2df7d4f2de7610b31518ee1573320566; expires=Sun, 08-Nov-20
-        17:29:26 GMT; path=/; domain=.nsone.net; HttpOnly
+      - __cfduid=d892cfccbfa1414b491fb85101d8db1f51573495252; expires=Tue, 10-Nov-20
+        18:00:52 GMT; path=/; domain=.nsone.net; HttpOnly
       X-Ratelimit-Remaining:
-      - '888'
+      - '895'
       Etag:
-      - W/"f9465a1545b8740fb4490490bfa9a823f7d729d4"
+      - W/"73f11d58a8f37980a21e6fbe0738aa8e729a3b5e"
       X-Ratelimit-By:
       - customer
       X-Ratelimit-Limit:
@@ -63,12 +63,12 @@ http_interactions:
       Server:
       - cloudflare
       Cf-Ray:
-      - 53318242be49ab7c-YYZ
+      - 53422b0fdd77ca98-YYZ
     body:
       encoding: ASCII-8BIT
-      string: '[{"nx_ttl":3600,"retry":7200,"dnssec":false,"zone":"dev.recordstore.io","network_pools":["p07"],"serial":1569855698,"primary":{"enabled":false,"secondaries":[]},"refresh":43200,"expiry":1209600,"dns_servers":["dns1.p07.nsone.net","dns2.p07.nsone.net","dns3.p07.nsone.net","dns4.p07.nsone.net"],"meta":{},"link":null,"primary_master":"dns1.p07.nsone.net","ttl":3600,"id":"5d921649bbccf90001ff95ee","hostmaster":"hostmaster@nsone.net","networks":[0],"pool":"p07"},{"nx_ttl":60,"retry":7200,"dnssec":false,"zone":"ns1.shopify.io","network_pools":["p03"],"serial":1570482853,"primary":{"enabled":false,"secondaries":[]},"refresh":43200,"expiry":1209600,"dns_servers":["dns1.p03.nsone.net","dns2.p03.nsone.net","dns3.p03.nsone.net","dns4.p03.nsone.net"],"meta":{},"link":null,"primary_master":"dns1.p03.nsone.net","ttl":60,"id":"5d715e23c94a900001192d45","hostmaster":"hostmaster@nsone.net","networks":[0],"pool":"p03"},{"nx_ttl":3600,"retry":7200,"dnssec":false,"zone":"test-for-chris.recordstore.com","network_pools":["p06"],"serial":1572893652,"primary":{"enabled":false,"secondaries":[]},"refresh":43200,"expiry":1209600,"dns_servers":["dns1.p06.nsone.net","dns2.p06.nsone.net","dns3.p06.nsone.net","dns4.p06.nsone.net"],"meta":{},"link":null,"primary_master":"dns1.p06.nsone.net","ttl":3600,"id":"5dc073d4403d5d0001e05cb8","hostmaster":"hostmaster@nsone.net","networks":[0],"pool":"p06"},{"nx_ttl":3600,"retry":7200,"dnssec":false,"zone":"test.recordstore.chrome","network_pools":["p02"],"serial":1569963467,"primary":{"enabled":false,"secondaries":[]},"refresh":43200,"expiry":1209600,"dns_servers":["dns1.p02.nsone.net","dns2.p02.nsone.net","dns3.p02.nsone.net","dns4.p02.nsone.net"],"meta":{},"link":null,"primary_master":"dns1.p02.nsone.net","ttl":3600,"id":"5d9250d5bbccf90001ffe1f8","hostmaster":"hostmaster@nsone.net","networks":[0],"pool":"p02"},{"nx_ttl":3600,"retry":7200,"dnssec":false,"zone":"test.recordstore.io","network_pools":["p07"],"serial":1573320566,"primary":{"enabled":false,"secondaries":[]},"refresh":43200,"expiry":1209600,"dns_servers":["dns1.p07.nsone.net","dns2.p07.nsone.net","dns3.p07.nsone.net","dns4.p07.nsone.net"],"meta":{},"link":null,"primary_master":"dns1.p07.nsone.net","ttl":3600,"id":"5dc6f750bbccf90001839ff7","hostmaster":"hostmaster@nsone.net","networks":[0],"pool":"p07"},{"nx_ttl":3600,"retry":7200,"dnssec":false,"zone":"test.yml.recordstore.io","network_pools":["p06"],"serial":1571861668,"primary":{"enabled":false,"secondaries":[]},"refresh":43200,"expiry":1209600,"dns_servers":["dns1.p06.nsone.net","dns2.p06.nsone.net","dns3.p06.nsone.net","dns4.p06.nsone.net"],"meta":{},"link":null,"primary_master":"dns1.p06.nsone.net","ttl":3600,"id":"5db0b4a4c9c79d000163dce4","hostmaster":"hostmaster@nsone.net","networks":[0],"pool":"p06"},{"nx_ttl":1800,"retry":600,"dnssec":false,"zone":"tm2.shopifysvc.com","network_pools":["p07"],"serial":1571155013,"primary":{"enabled":false,"secondaries":[]},"refresh":3600,"expiry":604800,"dns_servers":["dns1.p07.nsone.net","dns2.p07.nsone.net","dns3.p07.nsone.net","dns4.p07.nsone.net"],"meta":{},"link":null,"primary_master":"dns1.p07.nsone.net","ttl":900,"id":"5da5ec45c94a900001a5c936","hostmaster":"hostmaster@nsone.net","networks":[0],"pool":"p07"}]
+      string: '[{"nx_ttl":3600,"retry":7200,"dnssec":false,"zone":"dev.recordstore.io","network_pools":["p07"],"serial":1569855698,"primary":{"enabled":false,"secondaries":[]},"refresh":43200,"expiry":1209600,"dns_servers":["dns1.p07.nsone.net","dns2.p07.nsone.net","dns3.p07.nsone.net","dns4.p07.nsone.net"],"meta":{},"link":null,"primary_master":"dns1.p07.nsone.net","ttl":3600,"id":"5d921649bbccf90001ff95ee","hostmaster":"hostmaster@nsone.net","networks":[0],"pool":"p07"},{"nx_ttl":60,"retry":7200,"dnssec":false,"zone":"ns1.shopify.io","network_pools":["p03"],"serial":1570482853,"primary":{"enabled":false,"secondaries":[]},"refresh":43200,"expiry":1209600,"dns_servers":["dns1.p03.nsone.net","dns2.p03.nsone.net","dns3.p03.nsone.net","dns4.p03.nsone.net"],"meta":{},"link":null,"primary_master":"dns1.p03.nsone.net","ttl":60,"id":"5d715e23c94a900001192d45","hostmaster":"hostmaster@nsone.net","networks":[0],"pool":"p03"},{"nx_ttl":3600,"retry":7200,"dnssec":false,"zone":"test-for-chris.recordstore.com","network_pools":["p06"],"serial":1572893652,"primary":{"enabled":false,"secondaries":[]},"refresh":43200,"expiry":1209600,"dns_servers":["dns1.p06.nsone.net","dns2.p06.nsone.net","dns3.p06.nsone.net","dns4.p06.nsone.net"],"meta":{},"link":null,"primary_master":"dns1.p06.nsone.net","ttl":3600,"id":"5dc073d4403d5d0001e05cb8","hostmaster":"hostmaster@nsone.net","networks":[0],"pool":"p06"},{"nx_ttl":3600,"retry":7200,"dnssec":false,"zone":"test.recordstore.chrome","network_pools":["p02"],"serial":1569963467,"primary":{"enabled":false,"secondaries":[]},"refresh":43200,"expiry":1209600,"dns_servers":["dns1.p02.nsone.net","dns2.p02.nsone.net","dns3.p02.nsone.net","dns4.p02.nsone.net"],"meta":{},"link":null,"primary_master":"dns1.p02.nsone.net","ttl":3600,"id":"5d9250d5bbccf90001ffe1f8","hostmaster":"hostmaster@nsone.net","networks":[0],"pool":"p02"},{"nx_ttl":3600,"retry":7200,"dnssec":false,"zone":"test.recordstore.io","network_pools":["p04"],"serial":1573495248,"primary":{"enabled":false,"secondaries":[]},"refresh":43200,"expiry":1209600,"dns_servers":["dns1.p04.nsone.net","dns2.p04.nsone.net","dns3.p04.nsone.net","dns4.p04.nsone.net"],"meta":{},"link":null,"primary_master":"dns1.p04.nsone.net","ttl":3600,"id":"5dc9a15cc9c79d0001326a42","hostmaster":"hostmaster@nsone.net","networks":[0],"pool":"p04"},{"nx_ttl":3600,"retry":7200,"dnssec":false,"zone":"test.yml.recordstore.io","network_pools":["p06"],"serial":1571861668,"primary":{"enabled":false,"secondaries":[]},"refresh":43200,"expiry":1209600,"dns_servers":["dns1.p06.nsone.net","dns2.p06.nsone.net","dns3.p06.nsone.net","dns4.p06.nsone.net"],"meta":{},"link":null,"primary_master":"dns1.p06.nsone.net","ttl":3600,"id":"5db0b4a4c9c79d000163dce4","hostmaster":"hostmaster@nsone.net","networks":[0],"pool":"p06"},{"nx_ttl":1800,"retry":600,"dnssec":false,"zone":"tm2.shopifysvc.com","network_pools":["p07"],"serial":1571155013,"primary":{"enabled":false,"secondaries":[]},"refresh":3600,"expiry":604800,"dns_servers":["dns1.p07.nsone.net","dns2.p07.nsone.net","dns3.p07.nsone.net","dns4.p07.nsone.net"],"meta":{},"link":null,"primary_master":"dns1.p07.nsone.net","ttl":900,"id":"5da5ec45c94a900001a5c936","hostmaster":"hostmaster@nsone.net","networks":[0],"pool":"p07"}]
 
         '
     http_version: 
-  recorded_at: Sat, 09 Nov 2019 17:29:26 GMT
+  recorded_at: Mon, 11 Nov 2019 18:00:53 GMT
 recorded_with: VCR 5.0.0
